moved statmsg for items
