+++ pol-core/pol/statmsg.cpp
@@ -15,6 +15,7 @@
 #include "clib/rawtypes.h"
 #include "pol/globals/network.h"
 #include "pol/globals/settings.h"
+#include "pol/item/item.h"
 #include "pol/mobile/charactr.h"
 #include "pol/network/client.h"
 #include "pol/network/packethelper.h"
@@ -317,5 +318,24 @@ void send_update_hits_to_inrange( Mobile::Character* chr )
 
   // Exclude self... otherwise their status-window shows 1000 hp!! >_<
   transmit_to_others_inrange( chr, &msg->buffer, msg->offset );
+}
+
+void send_short_statmsg( Network::Client* client, Items::Item* item )
+{
+  Network::PktHelper::PacketOut<Network::PktOut_11> msg;
+  msg->offset += 2;  // msglen
+  msg->Write<u32>( item->serial_ext );
+  msg->Write( Clib::strUtf8ToCp1252( item->name() ).c_str(), 30, false );
+
+  msg->WriteFlipped<u16>( Clib::clamp_convert<u16>( item->hp_ * 1000 / item->maxhp() ) );
+  msg->WriteFlipped<u16>( 1000_u16 );
+  msg->Write<u8>( 0_u8 );
+  msg->Write<u8>( 0_u8 );  // moreinfo
+
+  u16 len = msg->offset;
+  msg->offset = 1;
+  msg->WriteFlipped<u16>( len );
+
+  msg.Send( client, len );
 }
 }  // namespace Pol::Core
